pyproject.toml
[project]
name = "simple-employee-manager"
version = "0.1.0"
description = "Warehouse Employee Management System - Desktop application for tracking employees, certifications, and compliance"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "peewee>=3.17.0",
    "python-dateutil>=2.8.0",
    "openpyxl>=3.1.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "tabulate>=0.9.0",
    "questionary>=2.0.0",
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "freezegun>=1.5.0",
    "ruff>=0.8.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=html",
    "--ignore=tests/test_ui",
    "--ignore=tests/test_main_window.py",
]

[tool.ruff]
target-version = "py314"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults (common in typer/cli)
    "B904",  # raise from in exception handling (too strict for now)
    "UP045", # Use X | None (optional, can enable later)
    "F403",  # from module import * (unable to detect undefined names)
    "F405",  # may be undefined from star imports
]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["F401", "F841", "E722", "F541", "I001"]
"src/cli/*.py" = ["B008", "B904", "UP045", "F401", "F841", "F541", "C416"]
"src/employee/models.py" = ["F403", "F405"]
"src/employee/calculations.py" = ["F841"]
"src/ui_ctk/*.py" = ["E722", "F841"]
"src/ui_ctk/**/*.py" = ["E722", "F841"]
"tests/conftest.py" = ["E722"]
"src/state/app_state.py" = ["E402"]
"src/excel_import/*.py" = ["F401"]
"src/export/*.py" = ["F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
fail_under = 40
